requirements.txt
openai
torch
PyPDF2
ollama
pyyaml
beautifulsoup4
lxml
python-dotenv
fastapi
uvicorn
python-multipart
psycopg2-binary
pgvector
docling
langgraph
